# Conflicts:
#   game-project/resources/CONFIG

+++ game-project/game/info3/game/modele/MoveableEntityClass/Scythe.java
@@ -6,7 +6,7 @@ public class Scythe extends Weapon{
 	
 	private static Scythe scythe;
 	public static int DAMAGE = 2;
-	public static double RANGE  = 5.0;
+	public static int RANGE  = 5;
 	public static double ALPHA  = 80.0;
 
 	private Scythe() {
